+++ ChangeLog.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added 
 
-- functions `str/nrest` and `str/butnlast`
+- functions `str/nrest`, `str/butnlast`, and `str/split-at`
 - module `:qrref`
 
 ### Fixed